move to sub directory and add init file
